Detect virtualization to avoid errors when trying to access soundcard, running pianoteq, etc.

+++ scripts/recipes/install_pianoteq_binary.sh
@@ -42,8 +42,10 @@ rm -rf $ZYNTHIAN_PLUGINS_DIR/lv2/Pianoteq*.lv2
 ln -s $ZYNTHIAN_SW_DIR/pianoteq6/$armdir/*.lv2 $ZYNTHIAN_PLUGINS_DIR/lv2
 
 # Generate LV2 presets (Pianoteq >= v7.2.1)
-version=$(./pianoteq --version | cut -d' ' -f4)
-if [[ $version > "7.2.0" ]]; then
-	rm -rf $ZYNTHIAN_MY_DATA_DIR/presets/lv2/"Pianoteq 7 "*"-factory-presets"*.lv2
-	./pianoteq --export-lv2-presets $ZYNTHIAN_MY_DATA_DIR/presets/lv2
+if [ "$(systemd-detect-virt)" == "none" ]; then
+	version=$(./pianoteq --version | cut -d' ' -f4)
+	if [[ $version > "7.2.0" ]]; then
+		rm -rf $ZYNTHIAN_MY_DATA_DIR/presets/lv2/"Pianoteq 7 "*"-factory-presets"*.lv2
+		./pianoteq --export-lv2-presets $ZYNTHIAN_MY_DATA_DIR/presets/lv2
+	fi
 fi
